requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.6.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
pgvector==0.2.4
psycopg2-binary==2.9.9

# For testing
aiohttp==3.9.1

# Document processing for RAG
PyPDF2==3.0.1
python-docx==1.1.0 